Fixed Yaml merging to not change types
